+++ src/views/LoginApp.vue
@@ -1,57 +1,22 @@
 <template>
   <h1>Let's Login</h1>
-  <table>
-    <tr>
-      <th>メールアドレス：</th>
-    </tr>
-    <tr>
-      <td><input type="email" v-model="email" /></td>
-    </tr>
-    <tr>
-      <th>パスワード：</th>
-    </tr>
-    <tr>
-      <td><input type="password" v-model="password" /></td>
-    </tr>
-  </table>
-  <button v-on:click="login">ログインする</button>
+  <MailLoginCom></MailLoginCom>
   <br />
-  <button v-on:click="googleLogin">googleでログイン</button>
+  <GoogleLoginCom></GoogleLoginCom>
 </template>
 
 <script>
-import {
-  getAuth,
-  signInWithEmailAndPassword,
-  signInWithRedirect,
-  GoogleAuthProvider,
-} from "firebase/auth"
+import MailLoginCom from "@/components/MailLoginCom.vue"
+import GoogleLoginCom from "@/components/GoogleLoginCom.vue"
 
 export default {
   data() {
-    return {
-      email: "",
-      password: "",
-    }
+    return {}
   },
-  methods: {
-    login() {
-      const auth = getAuth()
-      signInWithEmailAndPassword(auth, this.email, this.password)
-        .then(
-          //成功時の処理
-          alert("ログイン成功しました")
-        )
-        .catch(
-          //エラー時処理
-          alert("ログインできませんでした")
-        )
-    },
-    googleLogin() {
-      const auth = getAuth()
-      const provider = new GoogleAuthProvider()
-      signInWithRedirect(auth, provider)
-    },
+  methods: {},
+  components: {
+    MailLoginCom,
+    GoogleLoginCom,
   },
 }
 </script>
